+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add a MASE Summary ([#148](https://github.com/KIT-IAI/pyWATTS/issues/148))
 
 ### Changed
+* Retraining is triggered after all steps are transformed ([#117](https://github.com/KIT-IAI/pyWATTS/issues/117))
 
 ### Fixed
 * Try to handle non matching shapes in metrics. If not possible raise an exception.  ([#154](https://github.com/KIT-IAI/pyWATTS/issues/154))